+++ src/services/schema-file-updater.js
@@ -1,6 +1,7 @@
 const _ = require('lodash');
 const fs = require('fs');
 const logger = require('../services/logger');
+const { parameterize } = require('../utils/string');
 const { prettyPrint } = require('../utils/json');
 
 function SchemaFileUpdater(filename, collections, meta, serializerOptions) {
@@ -70,7 +71,7 @@ function SchemaFileUpdater(filename, collections, meta, serializerOptions) {
         action.type = null;
       }
 
-      setDefaultValueIfNecessary(action, 'endpoint', `/forest/actions/${_.kebabCase(action.name)}`);
+      setDefaultValueIfNecessary(action, 'endpoint', `/forest/actions/${parameterize(action.name)}`);
       setDefaultValueIfNecessary(action, 'httpMethod', 'POST');
       setDefaultValueIfNecessary(action, 'fields', []);
       setDefaultValueIfNecessary(action, 'redirect', null);