+++ src/utils/string.js
@@ -1,0 +1,3 @@
+const Inflector = require('inflected');
+
+exports.parameterize = value => (value ? Inflector.parameterize(value.trim()) : '');